#913 refactor: extract watchlist management
