+++ src/components/product/detail/ProductInfo.tsx
@@ -10,8 +10,8 @@ import "yet-another-react-lightbox/plugins/thumbnails.css";
 import { ProductImageGallery } from "@/components/product/detail/ProductImageGallery.tsx";
 import { useTranslation } from "react-i18next";
 import { ProductSharer } from "@/components/product/detail/ProductSharer.tsx";
-import { NotificationButton } from "@/components/product/buttons/NotificationButton.tsx";
-import { WatchlistButton } from "@/components/product/buttons/WatchlistButton.tsx";
+import { NotificationButton } from "@/features/watchlist/components/NotificationButton.tsx";
+import { WatchlistButton } from "@/features/watchlist/components/WatchlistButton.tsx";
 import { ProductPriceEstimate } from "@/components/product/detail/ProductPriceEstimate.tsx";
 import { H1 } from "@/components/typography/H1.tsx";
 import { Link } from "@tanstack/react-router";
