+++ src/features/watchlist/components/NotificationButton.tsx
@@ -1,6 +1,6 @@
 import { Button } from "@/components/ui/button.tsx";
 import { Bell, BellRing } from "lucide-react";
-import { useWatchlistNotificationMutation } from "@/hooks/watchlist/useWatchlistNotificationMutation.ts";
+import { useWatchlistNotificationMutation } from "@/features/watchlist/api/useWatchlistNotificationMutation.ts";
 import type { ComponentProps } from "react";
 import { cn } from "@/lib/utils.ts";
